+++ lib/src/upgrader.dart
@@ -3,6 +3,7 @@
  */
 
 import 'dart:async';
+import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';